refactor chatbox to put method and props directly on components
