+++ client/src/container/Chatbox.jsx
@@ -1,49 +1,29 @@
 import React, { Component } from 'react';
 import Messages from '../components/Messages';
 import Messagebox from '../components/Messagebox';
-import { connect } from 'react-redux';
+// import { connect } from 'react-redux';
 // import io from 'socket.io-client';
 
-class Chatbox extends Component {
-  constructor () {
-    super()
-    this.state = {
-      messages: [
-        {name: 'Yuyi', message: 'Hello, there'}, 
-        {name: 'Melanie', message: 'Hi Dad'}
-      ] 
-    }
-  }
-
-  componentWillMount() {
-    this.props.socket.on('connect', function(){
-      console.log('Connected on the client-side: chatbox');
-    });
-  }
-
-  updateMessage() {
-
-  }
-
+export default class Chatbox extends Component {
   render(){
     return (
       <div>
         <div className="chat_header">Chatbox</div>
-        <Messages messages={this.state.messages}/>
-        <Messagebox socket={this.props.socket} />
+        <Messages />
+        <Messagebox />
       </div>
       )
   }
 }
 
-function mapStateToProps(state) {
-  return {
-    peerId: state.PeerId.peerId, 
-    socket: state.Socket.socket
-  }
-}
+// function mapStateToProps(state) {
+//   return {
+//     peerId: state.PeerId.peerId, 
+//     socket: state.Socket.socket
+//   }
+// }
 
-export default connect(mapStateToProps)(Chatbox);
+// export default connect(mapStateToProps)(Chatbox);
